# Speed up lambda capture handling